Implement Exemplar-SVM querying (#2)
The original python prototype for scout used Exemplar SVMs, inspired by
Andrej Karpathy's
[tweet](https://twitter.com/karpathy/status/1647025230546886658)
highlighting their usefulness. This PR brings in this functionality to
querying in scout.

This port ended up being a bit of a bear in that SVM support in rust is
pretty limited; other libraries did not perform as well. I ultimately
(after trying `libsvm`) decided to use `liblinear` which is the same
library used in `scikit-learn`. I replicated the settings used in the
python library and was more or less able to get perfect replication of
the python PoC.